Local file playlist fix
Playlists with a local file as the last song entry present a Type Error due to null fields being iterated on line 283.  This type error was being caught and thrown on the catch on line 206. Adding a  ? to the feature variable on line 283 will ignore these null entries and successfully add the data that is present into the final CSV.